refactor: move DayOfWeek to render props components
