+++ src/@types/index.ts
@@ -1,13 +1,13 @@
 import { ComponentProps, ReactElement, ReactNode, SyntheticEvent } from 'react'
 
 import { Props as CalendarProps } from '../calendar/calendar'
-import { Props as DayOfWeekProps } from '../calendar/day_of_week'
 import { Props as DaysOfWeekProps } from '../calendar/days_of_week'
 import { Props as MonthHeaderProps } from '../calendar/month_header'
 import { Props as MonthProps } from '../calendar/month'
 import { Props as NoticeProps } from '../calendar/notice'
 import { Props as WeekProps } from '../calendar/week'
 import Day from '../RenderPropsComponents/Day'
+import DayOfWeek from '../RenderPropsComponents/DayOfWeek'
 
 export type INoticeType =
   | 'overlapping_with_disabled'
@@ -29,6 +29,7 @@ type IChildren = {
 }
 
 export type RenderPropsDay = (Props: ComponentProps<typeof Day>) => ReactElement
+export type RenderPropsDayOfWeek = (Props: ComponentProps<typeof DayOfWeek>) => ReactElement
 
 export type HandleOnDayClick = (
   event: SyntheticEvent<HTMLButtonElement>
@@ -43,25 +44,20 @@ export type OnDisabledDayClick = HandleOnDayClick
 /**
  * Render Props Interface
  */
-interface ICalendarRenderPropArgs extends CalendarProps, IChildren {}
+interface ICalendarRenderPropArgs extends CalendarProps, IChildren { }
 export type ICalendarRenderProp = (
   Props: ICalendarRenderPropArgs
 ) => ReactElement
 
-interface IMonthRenderPropsArgs extends MonthProps, IChildren {}
+interface IMonthRenderPropsArgs extends MonthProps, IChildren { }
 export type IMonthRenderProps = (Props: IMonthRenderPropsArgs) => ReactElement
 
-interface IWeekRenderPropsArgs extends WeekProps, IChildren {}
+interface IWeekRenderPropsArgs extends WeekProps, IChildren { }
 export type IWeekRenderProps = (Props: IWeekRenderPropsArgs) => ReactElement
 
-interface IDaysOfWeekRenderPropsArgs extends DaysOfWeekProps, IChildren {}
+interface IDaysOfWeekRenderPropsArgs extends DaysOfWeekProps, IChildren { }
 export type IDaysOfWeekRenderProps = (
   Props: IDaysOfWeekRenderPropsArgs
-) => ReactElement
-
-interface IDayOfWeekRenderPropsArgs extends DayOfWeekProps, IChildren {}
-export type IDayOfWeekRenderProps = (
-  Props: IDayOfWeekRenderPropsArgs
 ) => ReactElement
 
 interface IMonthHeaderRenderPropsArgs extends MonthHeaderProps {
@@ -74,5 +70,5 @@ export type IMonthHeaderRenderProps = (
   Props: IMonthHeaderRenderPropsArgs
 ) => ReactElement
 
-interface INoticeRenderPropsArgs extends NoticeProps, IChildren {}
+interface INoticeRenderPropsArgs extends NoticeProps, IChildren { }
 export type INoticeRenderProps = (Props: INoticeRenderPropsArgs) => ReactElement
