+++ src/RenderPropsComponents/DayOfWeek/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './DayOfWeek'
