+++ src/calendar/days_of_week.tsx
@@ -1,7 +1,6 @@
 import React, { FC } from 'react'
 
-import { IDayOfWeekRenderProps, IDaysOfWeekRenderProps } from '../@types'
-import DayOfWeek from './day_of_week'
+import { IDaysOfWeekRenderProps, RenderPropsDayOfWeek } from '../@types'
 
 const FRIDAY_INDEX = 4
 
@@ -9,7 +8,7 @@ export type Props = {
   blockClassName: string
   customRender?: IDaysOfWeekRenderProps
   daysOfWeek: string[]
-  renderDayOfWeek?: IDayOfWeekRenderProps
+  renderDayOfWeek: RenderPropsDayOfWeek
   weekStartsOn: number
 }
 
@@ -36,15 +35,12 @@ const DaysOfWeek: FC<Props> = props => {
   } = props
   const slicedDaysOfWeek = getDaysOfWeek({ dayIndex: weekStartsOn, daysOfWeek })
 
-  const children = slicedDaysOfWeek.map((day, index) => (
-    <DayOfWeek
-      blockClassName={blockClassName}
-      isWeekend={index > FRIDAY_INDEX}
-      day={day}
-      customRender={renderDayOfWeek}
-      key={day}
-    />
-  ))
+  const children = slicedDaysOfWeek.map((day, index) => renderDayOfWeek({
+    blockClassName,
+    day,
+    isWeekend: index > FRIDAY_INDEX,
+    key: day
+  }))
 
   if (customRender) {
     return customRender({
