+++ src/calendar/week.tsx
@@ -71,12 +71,12 @@ export default class Week extends Component<Props, {}> {
 
     return Boolean(
       selectedMin &&
-        selectedMax &&
-        isWithinRange(
-          startOfDay(date),
-          startOfDay(selectedMin),
-          startOfDay(selectedMax)
-        )
+      selectedMax &&
+      isWithinRange(
+        startOfDay(date),
+        startOfDay(selectedMin),
+        startOfDay(selectedMax)
+      )
     )
   }
 
@@ -150,9 +150,9 @@ export default class Week extends Component<Props, {}> {
         handleOnClick: isSelectable
           ? onDayClick
           : isDisabled
-          ? onDisabledDayClick
-          : () => {},
-        handleOnEnter: isSelectable ? onDayMouseEnter : () => {},
+            ? onDisabledDayClick
+            : () => { },
+        handleOnEnter: isSelectable ? onDayMouseEnter : () => { },
         isCurrentMonth,
         isDisabled,
         isHighlighted: this._dateHighlighted(day),
